+++ lua/telescope-docker/compose/picker.lua
@@ -1,8 +1,8 @@
-local mappings = require "telescope._extensions.docker.compose.mappings"
-local util = require "telescope._extensions.docker.util"
+local mappings = require "telescope-docker.compose.mappings"
+local util = require "telescope-docker.util"
 local builtin = require "telescope.builtin"
 local action_state = require "telescope.actions.state"
-local State = require "telescope._extensions.docker.util.docker_state"
+local State = require "telescope-docker.util.docker_state"
 
 local get_result_processor
 local name = "Docker compose files"
