+++ lua/telescope-docker/images/finder.lua
