+++ lua/telescope-docker/images/picker.lua
@@ -1,8 +1,8 @@
-local util = require "telescope._extensions.docker.util"
-local finder = require "telescope._extensions.docker.images.finder"
-local previewer = require "telescope._extensions.docker.images.previewer"
-local mappings = require "telescope._extensions.docker.images.mappings"
-local State = require "telescope._extensions.docker.util.docker_state"
+local util = require "telescope-docker.util"
+local finder = require "telescope-docker.images.finder"
+local previewer = require "telescope-docker.images.previewer"
+local mappings = require "telescope-docker.images.mappings"
+local State = require "telescope-docker.util.docker_state"
 
 local pickers = require "telescope.pickers"
 local conf = require("telescope.config").values
