+++ lua/telescope-docker/util/init.lua
@@ -1,4 +1,4 @@
-local enum = require "telescope._extensions.docker.enum"
+local enum = require "telescope-docker.enum"
 local telescope_actions = require "telescope.actions"
 
 local util = {}
