+++ lua/telescope/_extensions/docker.lua
@@ -1,10 +1,9 @@
-local util = require "telescope._extensions.docker.util"
-local setup = require "telescope._extensions.docker.setup"
-local containers_picker =
-  require "telescope._extensions.docker.containers.picker"
-local images_picker = require "telescope._extensions.docker.images.picker"
-local compose_picker = require "telescope._extensions.docker.compose.picker"
-local dockerfiles_picker = require "telescope._extensions.docker.files.picker"
+local util = require "telescope-docker.util"
+local setup = require "telescope-docker.setup"
+local containers_picker = require "telescope-docker.containers.picker"
+local images_picker = require "telescope-docker.images.picker"
+local compose_picker = require "telescope-docker.compose.picker"
+local dockerfiles_picker = require "telescope-docker.files.picker"
 
 -- NOTE: ensure the telescope is loaded
 -- before registering the extension
